move all PathUtility related code to one place
