+++ src/NuGet.Clients/NuGet.CommandLine/Commands/ProjectFactory.cs
@@ -22,7 +22,6 @@ using NuGet.ProjectManagement;
 using NuGet.Protocol;
 using NuGet.Protocol.Core.Types;
 using NuGet.Versioning;
-using PathUtility = NuGet.Commands.PathUtility;
 using XElementExtensions = NuGet.Packaging.XElementExtensions;
 
 namespace NuGet.CommandLine
