+++ src/NuGet.Clients/NuGet.CommandLine/Common/MSBuildProjectSystem.cs
@@ -112,7 +112,7 @@ namespace NuGet.Common
                 throw new ArgumentNullException(nameof(targetFullPath));
             }
 
-            var targetRelativePath = NuGet.Commands.PathUtility.GetRelativePath(PathUtility.EnsureTrailingSlash(ProjectFullPath), targetFullPath);
+            var targetRelativePath = PathUtility.GetRelativePath(PathUtility.EnsureTrailingSlash(ProjectFullPath), targetFullPath);
             var imports = Project.Xml.Imports;
             bool notImported = true;
             if (imports != null)
@@ -152,8 +152,8 @@ namespace NuGet.Common
 
         public void AddReference(string referencePath)
         {
-            string fullPath = NuGet.Commands.PathUtility.GetAbsolutePath(ProjectFullPath, referencePath);
-            string relativePath = NuGet.Commands.PathUtility.GetRelativePath(Project.FullPath, fullPath);
+            string fullPath = PathUtility.GetAbsolutePath(ProjectFullPath, referencePath);
+            string relativePath = PathUtility.GetRelativePath(Project.FullPath, fullPath);
             string assemblyFileName = Path.GetFileNameWithoutExtension(fullPath);
 
             try
@@ -274,7 +274,7 @@ namespace NuGet.Common
                 throw new ArgumentNullException(nameof(targetFullPath));
             }
 
-            var targetRelativePath = NuGet.Commands.PathUtility.GetRelativePath(PathUtility.EnsureTrailingSlash(ProjectFullPath), targetFullPath);
+            var targetRelativePath = PathUtility.GetRelativePath(PathUtility.EnsureTrailingSlash(ProjectFullPath), targetFullPath);
             if (Project.Xml.Imports != null)
             {
                 // search for this import statement and remove it
