+++ src/NuGet.Core/NuGet.Common/Strings.resx
@@ -123,6 +123,9 @@
   <data name="ArgumentNullOrEmpty" xml:space="preserve">
     <value>Argument cannot be null or empty</value>
   </data>
+  <data name="Argument_Must_Be_GreaterThanOrEqualTo" xml:space="preserve">
+    <value>Value must be greater than or equal to {0}</value>
+  </data>
   <data name="Error_FailedToCreateRandomFile" xml:space="preserve">
     <value>Failed to create random file for dotnet add pkg command.</value>
   </data>
