+++ src/NuGet.Core/NuGet.ProjectManagement/Utility/FileSystemUtility.cs
@@ -1,6 +1,7 @@
 ﻿// Copyright (c) .NET Foundation. All rights reserved.
 // Licensed under the Apache License, Version 2.0. See License.txt in the project root for license information.
 
+using NuGet.Common;
 using System;
 using System.Collections.Generic;
 using System.Diagnostics.CodeAnalysis;
